Update to most recent jetpack; warn about multiple creds.
The most recent addon-sdk now requires that "sdk/" be prefixed for SDK
modules, or at least it gets upset if you don't do the update.

The larger fix/warning is that I apparently had 2 sets of bugzilla
credentials on my Firefox.  One of which was out of date.  So depending
on the order in which the credentials were returned, things would either
fail or succeed.  Or maybe it would always fail on this profile and
succeed on another profile.  Unclear.

I didn't really fix the problem, but more have the extension stop arbitrarily
choosing a single credential and instead have it get upset.  This might not
be what you want.